requirements.txt
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
requests>=2.31.0
tenacity>=8.3.0
PyYAML>=6.0
colorama>=0.4.6
python-dotenv>=1.0.0
Levenshtein